updated floats to Decimals, updated versions of nuget packages

+++ NL/Models/Material.cs
@@ -14,8 +14,8 @@ namespace NL.Models
         [Key]
         public int MaterialID { get; set; }
         public String Name { get; set; }
-        public float Cost_per_g { get; set; }
-        public float Available_g { get; set; }
-        public float ReorderPoint_g { get; set; }
+        public Decimal Cost_per_g { get; set; }
+        public Decimal Available_g { get; set; }
+        public Decimal ReorderPoint_g { get; set; }
     }
 }
